kakao login refactoring and DebugLog setting
