+++ app/src/main/java/com/eroom/erooja/feature/login/LoginContract.kt
@@ -1,11 +1,15 @@
 package com.eroom.erooja.feature.login
 
+import com.kakao.usermgmt.response.MeV2Response
+import java.util.concurrent.Future
+
 interface LoginContract {
     interface View {
-
+        val redirectSignUpActivity: () -> Unit
     }
 
     interface Presenter {
         val view: View
+        val requestMe: () -> Future<MeV2Response>
     }
 }
