+++ app/src/main/java/com/eroom/erooja/feature/signup/kakao/KakaoSignUpActivity.kt
@@ -2,17 +2,7 @@ package com.eroom.erooja.feature.signup.kakao
 
 import android.os.Bundle
 import androidx.appcompat.app.AppCompatActivity
-import com.eroom.domain.utils.toastLong
 import com.eroom.erooja.R
-import com.kakao.network.ErrorResult
-import com.kakao.usermgmt.UserManagement
-import com.kakao.usermgmt.callback.LogoutResponseCallback
-import com.kakao.usermgmt.callback.MeV2ResponseCallback
-import com.kakao.usermgmt.response.MeV2Response
-import com.kakao.usermgmt.response.model.Profile
-import com.kakao.usermgmt.response.model.UserAccount
-import com.kakao.util.OptionalBoolean
-import timber.log.Timber
 
 
 class KakaoSignUpActivity : AppCompatActivity() {
@@ -21,9 +11,5 @@ class KakaoSignUpActivity : AppCompatActivity() {
         super.onCreate(savedInstanceState)
         setContentView(R.layout.activity_kakao_sign_up)
 
-
     }
-
-
-
 }
